Additional tests and move translators into subpackage.
