+++ flink-libraries/flink-gelly/src/test/java/org/apache/flink/graph/asm/translate/translators/LongValueToSignedIntValueTest.java
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package org.apache.flink.graph.asm.translate;
+package org.apache.flink.graph.asm.translate.translators;
 
+import org.apache.flink.graph.asm.translate.TranslateFunction;
 import org.apache.flink.types.IntValue;
 import org.apache.flink.types.LongValue;
 import org.junit.Test;
